iPhone X support.
The biggest complication was that the indicator height is no longer fixed so the various objects that depend on it need access back to the actual indicator.  There was also an issue with the content inset behavior in the draft selector.  Eventually, I should adopt the rounded corners and affordances of iOS 11, but I'm not there yet.